requirements.txt
google-adk
google-cloud-logging
google-genai
langchain
langchain-community
wikipedia
python-dotenv